requirements.txt
streamlit
openai
python-pptx
python-docx
openpyxl
PyPDF2
